+++ src/measureDistances.js
@@ -78,6 +78,59 @@ class AAmeasure {
         return false
     }
 
+    static isTokenInside(templateDetails, token, wallsBlockTargeting = false) {
+        const grid = canvas?.scene?.data.grid;
+        if (!grid)
+            return false;
+        const templatePos = { x: templateDetails.x, y: templateDetails.y };
+        // Check for center of  each square the token uses.
+        // e.g. for large tokens all 4 squares
+        const startX = token.data.width >= 1 ? 0.5 : (token.data.width / 2);
+        const startY = token.data.height >= 1 ? 0.5 : (token.data.height / 2);
+        for (let x = startX; x < token.data.width; x++) {
+            for (let y = startY; y < token.data.height; y++) {
+                const currGrid = {
+                    x: token.data.x + x * grid - templatePos.x,
+                    y: token.data.y + y * grid - templatePos.y,
+                };
+                let contains = templateDetails.shape?.contains(currGrid.x, currGrid.y);
+                if (contains && wallsBlockTargeting) {
+                    let tx = templatePos.x;
+                    let ty = templatePos.y;
+                    if (templateDetails.shape.type === 1) { // A rectangle
+                        tx = tx + templateDetails.shape.width / 2;
+                        ty = ty + templateDetails.shape.height / 2;
+                    }
+                    const r = new Ray({ x: tx, y: ty }, { x: currGrid.x + templatePos.x, y: currGrid.y + templatePos.y });
+                    if (configSettings.optionalRules.wallsBlockRange === "centerLevels" && installedModules.get("levels")) {
+                        let p1 = {
+                            x: currGrid.x + templatePos.x, y: currGrid.y + templatePos.y,
+                            //@ts-ignore
+                            z: token.data.elevation
+                        };
+                        const p2z = installedModules.get("levels")?.lastTokenForTemplate.data.elevation
+                            ?? installedModules.get("levels")?.nextTemplateHeight ?? 0;
+                        let p2 = {
+                            x: tx, y: ty,
+                            //@ts-ignore
+                            z: p2z
+                        };
+                        contains = getUnitDist(p2.x, p2.y, p2.z, token) <= templateDetails.distance;
+                        //@ts-ignore
+                        contains = contains && !installedModules.get("levels").testCollision(p1, p2, "collision");
+                        //@ts-ignore
+                    }
+                    else {
+                        contains = !canvas?.walls?.checkCollision(r);
+                    }
+                }
+                // Check the distance from origin.
+                if (contains)
+                    return true;
+            }
+        }
+        return false;
+    }
     static heightCheck(source, target, radius, r) {
         let distance;
         switch (game.settings.get("ActiveAuras", "vertical-euclidean")) {